Fix context-aware metrics server.
See https://github.com/golang/go/issues/24047. Apparently, it is
intentional that Go HTTP servers can be reused after they have been
terminated. This caused a race condition in the goroutine that checked
for context completion, as it could called Server.Shutdown() on a server
that had not even started yet. Now we just kill the underlying listener
instead.