New: Add secrets/encoding package to standardize secret value encoding for storage
To maintain secret value integrity, we need to encode the value bytes in a common encoding for
storage. We also need to make it transparent to the user so we don't force them to encode their
values before adding them to the secret store. This package also helps maintain backwards
compatibility with the current system that doesn't have any consistent value encoding.