Breaking: Revise scheduler to correctly propagate error behavior
This change represents effectively a V2 for the scheduler package.
Although the previous implementation worked sufficiently well, it had
serious limitations in its ability to manage errors (for example, a
Parent would exit as soon as any Segment exited, regardless of the
desired error behavior of the parent).

Additionally, the logic in the Segment implementation was intermixed
with the base functionality of managing Goroutines. As part of providing
the same error handling support to both Parent and Segment, it became
increasingly evident that there existed a common base between the two
lifecycles, namely, the ability to wait for a Goroutine to complete and
collect errors from it as it runs. This functionality is now exposed as
the Scheduler type.

Work in another project exposed the opportunity to provide an additional
Descriptor implementation to allow external parties to submit new work
to the scheduler, now provided as AdhocDescriptor and AdhocSubmitter.
This gives us rough parity to Java's ExecutorService.

We add documentation for most types and a few additional tests.