metrics: removes delegate field from prometheus timer metric
We often need the ability to update label values on a timer that's in
flight and originally we would create a delegate, that was a prometheus
Observer when the timer is started. This isn't needed as we can just
create the observer on the fly when we want to stop the timer and
observe the duration, allowing us to set the labels that can be passed
in.