Update: Add noop metrics delegate for testing
This commit adds a noop delegate to instrumentation/metrics that will just do nothing. It will not
register metrics or record them. This is used for testing packages that use Horsehead for
instrumenetation. Prometheus registers everything in a global registry at the moment, so errors can
arise in the lifecycle of tests related to metrics already having been registered. This can create
issues when inspecting the call stack and output logs of tests.