Add support for sending fake clocks in the context
This change makes it possible to expose a clock in the context. It also
provides utility methods to use such a clock in the same form as the
standard library context package, i.e., WithTimeout and WithDeadline now
respect fake clocks.

Code that could take advantage of passing a clock through the context
has been updated to do so.